+++ ReleaseNotes.md
@@ -1,3 +1,6 @@
+## 1.1.0 Named lines
++ Added Named lines support
+
 ## 1.0.1 Fix initial configuration
  + Fixed default values of initial configuration